moved HideOnScroll to its own component
