+++ src/frontend/src/components/HideOnScroll/index.js
@@ -1,0 +1,3 @@
+import HideOnScroll from './HideOnScroll.jsx';
+
+export default HideOnScroll;
